+++ packages/react/src/components/TextField/index.story.tsx
@@ -3,7 +3,11 @@ import React from 'react'
 import styled from 'styled-components'
 import { Story } from '../../_lib/compat'
 import Clickable from '../Clickable'
-import TextField, { TextFieldProps } from '.'
+import TextField, {
+  MultiLineTextFieldProps,
+  SingleLineTextFieldProps,
+  TextFieldProps,
+} from '.'
 import { px } from '@charcoal-ui/utils'
 import IconButton from '../IconButton'
 
@@ -34,8 +38,19 @@ const Template: Story<Partial<TextFieldProps>> = (args) => (
       subLabel={
         <Clickable onClick={action('label-click')}>Text Link</Clickable>
       }
-      placeholder="TextField"
-      {...args}
+      placeholder="Single Line"
+      {...(args as Partial<SingleLineTextFieldProps>)}
+      multiline={false}
+    />
+    <TextField
+      label="Label"
+      requiredText="*必須"
+      subLabel={
+        <Clickable onClick={action('label-click')}>Text Link</Clickable>
+      }
+      placeholder="Multi Line"
+      {...(args as Partial<MultiLineTextFieldProps>)}
+      multiline
     />
   </Container>
 )
@@ -55,7 +70,7 @@ HasCount.args = {
   maxLength: 100,
 }
 
-export const HasAffix: Story<Partial<TextFieldProps>> = (args) => (
+export const HasAffix: Story<Partial<SingleLineTextFieldProps>> = (args) => (
   <TextField label="Label" placeholder="path/to/your/file" {...args} />
 )
 HasAffix.args = {
@@ -65,7 +80,14 @@ HasAffix.args = {
   suffix: '.png',
 }
 
-export const PrefixIcon: Story<Partial<TextFieldProps>> = (args) => (
+export const AutoHeight: Story<Partial<MultiLineTextFieldProps>> = (args) => (
+  <TextField label="Label" placeholder="Multi Line" {...args} multiline />
+)
+AutoHeight.args = {
+  autoHeight: true,
+}
+
+export const PrefixIcon: Story<Partial<SingleLineTextFieldProps>> = (args) => (
   <TextField
     label="Label"
     placeholder="Icon prefix"
